requirements.txt
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
moviepy==2.1.2
pyttsx3==2.90
numpy<2.0.0  # Pour éviter les problèmes de compatibilité
mistralai==0.0.12
pydantic==2.6.1
boto3==1.34.34  # Pour AWS S3
google-cloud-storage==2.14.0  # Pour Google Drive
pytest==8.0.0  # Pour les tests 